Make our SQLiteQueryBuilder even stronger.
As part of making MediaProvider into a Mainline module, we need to
strictly enforce the APIs that we support.  Last year we did this
for incoming query() calls via setProjectionMap(), and this change
completes that work for update() and delete() calls.

This change also enables strict grammar checking to prevent
subqueries which can result in a slow side-channel attack that
can leak otherwise hidden information.

Bug: 137890034
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: If5f552912f624efa55b89bbeb3ea89e8168d7825