Don't redact if caller is owner and add redaction test
Add check for owner package name when redacting a file, if the owner is
reading the file or of the reader has been granted write permission to
the corresponding URI, we don't redact.

Also add tests for redaction, however, no test for the URI permission
grant has been added, yet. Tracked in b/146346138.

Fixes: 145795132
Test: atest FuseDaemonHostTest

Change-Id: Ibf5900c53d27af9f17204f8f294f915c81b04614