FuseDaemon: Remove bogus null checks after new.
By default, new will throw std::bad_alloc (and bring down the process)
if it fails, it will never return null unless "new (std::nothrow)" is
used but that's not really advisable as there's rarely anything to be
gained by being resilient to memory exhaustion.

This is step one of several in getting rid of our explicit setting of
errno.

Test: atest FuseDaemonHostTest
Change-Id: Ia7e2ad2088eaea979934e1cba5316c38e2111304