Only reconcile media added before scan started.
If an app inserts a media item after scan finishes with a particular
directory, we may not include it in mScannedIds, and we'd then end up
treating that newly insert item as an mUnknownIds and delete it.

Avoid this by only considering items for mUnknownIds that were added
before the scan started.

Bug: 146167762
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: Ia677fd3904f8c46f748370a975d3fa3f4583f4f7