Adjust database views based on mounted volumes.
This is a continuation of the work started in cc5c31ddb, which
adjusted the IDs of artists, albums, and genres to be based on hashes
which allows us to quickly identify values that should be clustered,
regardless of which volume name the underlying media is stored on.

We now dynamically adjust the artists, albums, and genres views based
on which storage volumes are currently mounted on the device.  Tests
verify this works as expected as we change which volumes are mounted.

Add warning note about golden schemas, since they should never be
manually changed.

Bug: 143722813
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I0d4f75e3b8ff2a2b8f3cdf389c4b69da33dde16c