Prefer earlier-sorted strings for tie-breaking.
When a user has media with both "Beatles" and "beatles", let's use
the MIN() function to prefer selection of "Beatles", as it begins
with capital letters.

Once the platform adds SQLiteDatabase.addCustomFunction() as public
API in a future release we can improve this by implementing a more
granular tie-breaking algorithm.

Bug: 137395060
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: Iebc4d3ca60658fe291ef7c79f7dd5fd73e2aef7e