Return normal (non-negated) errno to FuseDaemon
Previously, MediaProvider used to return negated errno values in the
case of a failure. This was needed because the open operation used to
happen in MediaProvider and we needed a way to distinguish error from
innocent FDs. Now that open is done in FuseDaemon, we can return normal
errno values instead.

Test: atest FuseDaemonHostTest
Fixes: 145914187
Change-Id: I03a4c29c875c243ea4eb6171d1d7e6b33ed6f91a