Reconcile directory and file paths correctly in sql query for scan
Previously, for a deleted file, scan() wouldn't remove corresponding
database entry. The SQLite query used in scan logic used path of file
appended with '/%' for matching corresponding entry in MediaProvider
database. We append '/%' for directory paths i.e., when mSingleFile is
false, but mSingleFile can be false for deleted files also.
New logic doesn't rely on mSingleFile, instead tries to match directory
path with '/%' and file path as it is using 'OR' in sqlite query to
match both cases right.

Test: atest --test-mapping packages/providers/MediaProvider
Test: atest CtsDownloadManagerApi28
Bug: 147595996
Change-Id: Id44baafd7e97042af02970e3fd8a2602005f9fba