Add app cache clearing mechanism
Adds a dialog, implemented as an activity in MediaProvider, to prompt
the user to clear app cache directories. The dialog is triggered by
starting an activity with the intent action
StorageManager.ACTION_CLEAR_APP_CACHE. The activity will be actually
displayed only if the calling package has MANAGE_EXTERNAL_STORAGE
permission or OP_MANAGE_EXTERNAL_STORAGE allowed.

The activity will return Activity#RESULT_OK if the user allowed the
cache dirs to cleared, and Activity#RESULT_CANCELED in every other case,
including: if the caller doesn't have the right permissions or if the
user decided not to clear app caches.

Bug: 144341120
Test: manual

Change-Id: I2409cc3ad44ae58f1979f23e53adc92814b81b96