Check updated volume cache before failing to find volume
At system boot, the StorageManagerService broadcasts when a volume is
mounted. The MediaProvider listens for these callbacks/broadcasts to
update it's cache of volumes it knows about.

There may be a race where the MediaProvider gets a request for a
volume before it's cache is updated, causing that request to fail with
a VolumeNotFoundException

We now update the cache and check the latest cache before failing.

This seems to have been the cause of some flaky presubmit tests

Test: atest FuseDaemonHostTest
Bug: 149132662
Change-Id: Ie07ed991d02473713054d5879ec3d7b1d104c863