Save calling UID on FUSE pf_create.
File is inserted into databse on FUSE RELEASE operation using scanFile.
This insert does not add owner package name. And, caller of RELEASE may
not be the owner of the file. To know the owner of the file, save the
calling UID information on create() file. This UID can be used in
scanFileForFuse() to update owner information in the database.

Test: atest FuseDaemonHostTest
Test: atest fuse_node_test
Change-Id: Ibc7f43e3f2cb2c76331206cbfc2d82f09277e712