Set filesystem quota type when scanning files.
Devices launching with R use project-ID based filesystem quotas, meaning
that each file is tagged with an associated ID that describes what type
of file it is. Previously sdcardfs in the kernel took care of this, but
now that sdcardfs is going away, have MediaProvider setup the type
correctly - it indexes all the files on external storage anyway, so it
is the best place to do it.

Bug: 146419093
Test: manual inspection with lsattr -p -R
Change-Id: I3e0050c99849fefee0bc0045d8ad63483e6d74ce