FuseDaemon: Update JNI threading model.
JNI up-calls are now performed on the same thread on which
they're required (usually a libfuse thread) and not on a
separate dedicated thread. We attach each thread the first
time we make a call on it and detach it when the thread is
about to be destroyed.

Test: atest FuseDaemonHostTest
Bug: 149198252

Merged-In: I04795d069e8e5c3ac2d572dfc48a7e2146bc31c5
Change-Id: I04795d069e8e5c3ac2d572dfc48a7e2146bc31c5
(cherry picked from commit 402a85fa39e9f8b12fc9967ceb7ca52c3db35b41)