Invalidate FUSE VFS dentry cache
To improve filesystem performance we can reduce unnecessary
requests to the FUSE daemon by enabling VFS caching of dentries
after a FUSE_LOOKUP request.

To enable correctly, we should ensure that the lower filesystem
dentries are not modified outside of the FUSE driver. Unfortunately,
we already do this when handling IO requests over binder using the
ContentResolver interface.

To fix, we should invalidate any FUSE dentries in the VFS cache so that
subsequent lookups for those files do not return incorrectly cached
values from the VFS.

This change just adds support for invalidating FUSE VFS cache
dentries, subsequent cl will call it when required. Note that it
should never be called in the execution path of a related filesytem
operation. See fuse_lowlevel.h for more details. Will ensure that this
is never called in the FUSE execution path in a follow up

Test: atest FuseDaemonHostTest#testVfsCacheConsistency
Bug: 145741152
Change-Id: I54e67f63a29c82a392916fb28d1a7f8f894c685b