Remove stale old path from database on file rename
For apps that bypass scoped storage, file/directory is renamed in the
lower file system. Database rows corresponding to these files are not
updated with new path. This leaves database in an inconsistent state
with stale entries of files that are renamed to new path. Adding a
workaround to delete old path from database using scanFile.

Test: On video capture from camera app, old path of the renamed file is
deleted from database.
Test: atest FuseDaemonHostTest
Bug: 149661345

Change-Id: I4a774b4d332de94c3f23e8cd466312712ad50f96