Check for READ_EXTERNAL_STORAGE when querying
MediaProvider#getQueryBuilder would give special privileges to legacy
apps, but it always checks for WRITE_EXTERNAL_STORAGE, even if the
caller is just querying.

This change makes getQueryBuilder check for READ_EXTERNAL_STORAGE when
the caller is querying, and WRITE_EXTERNAL_STORAGE when inserting,
deleting or updating.

This is a step towards ensuring we noteOp the relevant app-ops in
MediaProvider.

Test: atest --test-mapping]
Bug: 149473595
Change-Id: I65fae17bd706bbab614c571184fdf0c7157479cb