Revert create/update/delete database for apps that bypass restrictions.
Apps are not aware of FuseDaemon modifying the database. Apps that
assume the old uri to be available even after rename/delete don't
work correctly because of the new behaviour.
As a temporary fix for this issue, reverting create/update/delete
database row from FuseDaemon for apps that bypass scoped storage
restrictions. Long term solution will be added as part of b/145737191.

Ignoring some FuseDaemonHostTest that fail becuase of the revert.

Test: atest FuseDaemonHostTest --iterations 2
Bug: 150147690
Change-Id: Ia0a8fe2461133ed0ef68384c4856c0fc93735541