Deny apps with broad access from accessing other apps' dirs
Even apps with broad access to storage like file managers or galleries
should not be able to access other apps' external storage directories.
This is mostly guaranteed by bind-mounting Android/data and Android/obb,
but this is not true for e.g. removable volumes, so we need to enforce
that restriction in MediaProvider as well.

In addition, this gives us more control over allowing access to
Android/media dirs.

Test: atest FuseDaemonHostTest
Bug: 150467040
Change-Id: Ifba6d62ce65a30058cd337e462b5903c0de2b3aa