Create test directory with the correct permissions.
The /sdcard/Android/data/com.android.shell directory is created by root,
which has a umask of '000', and therefore creates the directory with
permissions of 777, making the directory world-writable. Before,
sdcardfs magically fixed up the permissions, but now that sdcardfs is
removed on cuttlefish, that is no longer the case.

Ideally this test should just have an app itself create it's own
app-private data dir instead of manually creating it, but to unblock
flakiness, manually set the correct permissions on this directory.

Bug: 150467040
Test: atest
com.android.tests.fused.host.LegacyAccessHostTest#testCreateFilesInRandomPlaces_hasW
Change-Id: If7cb02d80ec981a4baf9dfe7da711de7000fc3d4