Treat playlists like first class media files
Playlists, like images or audio, can (and often do) have a corresponding
file on the file system. For that reason, they need to be treated like
media files by MediaStore. This means that their db view should inherit
all of the columns from MediaColumns.
In addition, this allows us to query for OWNER_PACKAGE_NAME of a
playlist or open a playlist file through ContentResolver, which we
previously couldn't do (unless we go to the files table).

Test:  * Uri item = getContentResolver().insert(
                 MediaStore.Audio.Playlists.EXTERNAL_CONTENT_URI, values);
       * getContentResolver().openFileDescriptor(item, "rw");
Bug: 150593090

Change-Id: I42a87358978053f8aacf92c363188762e9e4fadb