Fix VFS cache inconsistency in pf_create
The VFS cache fixes in pf_open should be added to pf_create because
there is a chance that a file is already created when pf_create is
executed. In that case we'd do an open (if the app didn't specify
O_EXCL).

Note that for pf_create to be called, the following should have
happened:
1. App calls open("foo", O_CREAT)
2. Kernel sends FUSE_LOOKUP to FuseDaemon
3. FuseDaemon responds ENOENT to lookup
4. pf_create

Between 3 and 4, 'foo' can be created on the lower file system over
binder. This can lead to the same issues we tried fixing in
I94c8cbed32567e807259d67a6a9a4ea470158d9e

Test: atest FuseDaemonHostTest
Bug: 148709965

Change-Id: I88f753861eaf21c05cf149b40d104e653428fb93