Don't run with a umask of 0.
On devices with sdcardfs, sdcardfs fixing up the permissions; but on
devices without sdcardfs, setting the umask to 0 results in all
files/dirs creates from MediaProvider to have mode 666/777 - this is too
broad, and instead we should explicitly grant more broad permissions
when we need to. There currently doesn't seem to be a need, because only
MediaProvider needs to be able to access these files.

Bug: 151044530
Test: TEST_MAPPING for this dir
Change-Id: I2ace8e24210f90f9deda7c48f55b57541a307734