Prevent invalidating dentries on FUSE threads
libfuse provides an api to invalidate VFS dentries,
fuse_lowlevel_notify_inval_entry. This can be useful to invalidate a
dentry if it has been modified on the lower file system.

This API *must* never be called on a FUSE thread.

This cl takes advantage of the fact that all FUSE threads can be
identified by a JNIEnv TLS pthread key.

Test: Manual
Bug: 149512775
Change-Id: Idc9b91d1979489568c4b1dea57212c5983950b52