Support checking FUSE thread execution
There are several cases where we want to behave differently depending
on if we are handling a FUSE request or Binder request. This cl adds a
method to support this check.

It relies on a TLS variable that would only be set if running on a
FUSE thread.

Test: m
Bug: 149512775
Merged-In: I8f77470443e92812598f0e90ca42b43ef25b9d10
Change-Id: I8f77470443e92812598f0e90ca42b43ef25b9d10
(cherry picked from commit 0602a41451b40264c08f48e772b16ac9d831d83c)