Allow ".mp4" files to masquerade as ".m4a".
We've heard from some partners that an online music service has been
delivering music files using the ".mp4" file extension, when they
should have been using the ".m4a" file extension.

We're normally very picky about only allowing MediaMetadataRetriever
to override a MIME type when the top-level of the type matches with
the file extension, to avoid chaos like a ".jpg" file pretending to
be an audio file, but we're willing to make this very narrow
concession.  We continue to feel *very* strongly that the top-level
of a MIME type should remain matching.

Bug: 148316354
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I3091f5dd022672c19d1e8fbe1835504a61beb3c0