Only decode full-bleed images when supported.
When we have a file format that we can't decode with ImageDecoder,
fall back to using the thumbnail as a stand-in.  This fixes a crash
when showing a full-bleed representation of video and audio files.

Also optimize full-bleed decoding based on screen size, since
we'll never need resolution beyond that.

Bug: 147883786
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I3733a6e7f26d1fa4802bab73d9c4fd804a3f45fb