Add sqlite3 debug scripts
Running sqlite3 CLI on the device can be tedious to use because it
doesn't allow repeating/editing commands or history.

This change adds 2 useful functions:
1. sqlite3-pull: Pulls db along with WAL journal file to get the
latest rows that might not have been committed to the db file.
Additionally, this commands deletes the insert/update/delete triggers
from the pulled db. Otherwise, the corresponding db operations would
fail because the requisite _insert/_update/_delete functions would be
unavailable

2. sqlite3-push: Pushes the (possibly modified) db and WAL journal
file to the device. Additionally trigering the db update flow so that
triggers can be recreated

Test: Manual
Bug: 151076202
Change-Id: Id8f218310ae5a41400812b42e4bd0db89aff58d4