Remove duplicate MediaStore class
This was originally added for MediaProvider to be able to
access hidden APIs in MediaStore. There's now a better approach
of doing this, by compiling against module_current, which
doesn't contain any module apis, and listing the module dependencies
explicitly. MediaProvider links against the stubs of statsd and
the impl jar of framework-mediaprovider.

MediaProvider is an apk-in-apex, so is whitelisted to access
blacklisted APIs.

Note that the sources of framework-mediaprovider are still linked in
to the MediaProviderTests, which aren't whitelisted. We'd like to have
a general solution for android_test rules to get whitelisted.

Bug: 152891096
Test: http://b/146894977#comment15
Change-Id: I7e2bc7475d47e68517f612b94585d00635cff747