Remove JNI from tests while remaining green.
The MTS infrastructure seems to be confused about which ABI to select
when installing tests that include native code.  The good news is
that we technically don't need JNI in our tests; we simply need to
replace it with an override that indicates that we're never running
a FUSE daemon.

Bug: 151763985
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I2ed699c8b3b9e1f3675c8b3ffb829007fbdbe729