Workaround to enable screenshots.
Movement of files on disk is normally only offered to "real" apps,
since system internals (such as ModernMediaScanner) never intend to
actually move files; they're simply updating things in place.

This approach works fine for most apps, but unfortunately SystemUI
is still being treated as isCallingPackageSystem() because the
work in 151768142 hasn't landed yet.  We temporarily work around
this with the smallest possible patch to avoid any additional
regressions, and this can be reverted as part of fixing 151768142.

Bug: 153209299, 151768142
Test: manual
Change-Id: I058e9593b44be801e63a9e092ca79c897edb228f