Invalidate VFS dentry when opening lower filesystem fd
With write back (wb) cache enabled, the kernel ignores some changes on
the lower filesystem if it has a dentry in it's cache.

Now, we invalidate the FUSE dentry before we return a writable fd to
the lower filesystem from MediaProvider.

This allows subsequent open and getattr calls to the FUSE daemon to
work correctly.

Test: atest FuseDaemonHostTest#testVfsCacheConsistency
Bug: 145741852
Change-Id: Ie0ea946245fe67e8ad93e3da4f59d6b87d103331