Don't overwrite is_download column
We set is_download=true when Downloads uri is used in
ContentResolver#insert. We changed computeValuesFromData to overwrite
this value and compute is_download column value based on the path.
Files downloaded by DownloadManager and not saved in Download directory
are also treated as downloaded files. Hence, we will not be able to set
is_download column based on the path of the file.

Removed overwriting of is_download column based on the path of the
downloaded file.

Test: atest CtsDownloadManagerApi28
Bug: 150672865, 153325386

Change-Id: Ifaed9d9b71c22b1f010baa21b93468a3940f15b0