Don't rewrite hidden file name for FUSE
While inserting a new row for hidden files, we change hidden file to
non-hidden by appending '_' to the file name. This ensures the row
corresponding to hidden file is not deleted from the database on a scan.

We should never rewrite file name for file path operations. Changed
ensureFileColumns to not rewrite hidden file name for FUSE operations.

Test: atest FuseDaemonHostTest#testCanCreateHiddenFile
Test: atest FileUtilsTest#testSanitizeValues_HiddenFile
Bug: 148585977
Change-Id: I6cb833eb4c9c40cc89bfc8566137aefd9873f5fc