Allow file movement via PermissionActivity.
Recent changes to trashing/untrashing items require it to be applied
through the "movement" code path of updateInternal().  This works
fine when apps are updating values they own directly, but it breaks
when called indirectly via PermissionActivity, since we skip the
movement behavior entirely for all MediaProvider internals.

This change introduces a QUERY_ARG_ALLOW_MOVEMENT that can be used
to force-enable the movement feature, fixing the broken tests.

Bug: 154354407, 151979369
Test: atest CtsAppSecurityHostTestCases:android.appsecurity.cts.ExternalStorageHostTest#testMediaEscalation
Change-Id: I472cd2b66df93a0d2a6d22edcee365ad0529ceeb